adds OptionsCore trait and proc macro
OptionsCore is a subset of Options. it removes most of the command listing/ help printing/default behavior like exiting, etc. from the trait. Additionally it alters the parsing behavior to not treat -h/--help as a special variable, dont use the first letter of a field as a short option (user must do that manually with #[option(short = 'h')] ), and does not treat unknown input as errors.